Add files via upload
Rainbows are visually striking manifestations of refraction and reflection. Model describes the formation of rainbows and explores how the shape, size and constitution
of the liquid droplets affect them. In the first instance, model aims to give
quantitative results for the angle dependence of the intensity of different wavelengths.
There many further questions that have been answered, including:
• How are secondary rainbows created?
• Can you reproduce the ‘Alexander’s Band’?
• As droplets fall through the air, their shape can deviate from a perfect sphere. How
does this affect the shape of the rainbow?
• How does the refractive index of the liquid affect the rainbow?
• How does the polarization of the light affect the bow, and does the bow affect the
polarization of the rainbow in turn?
• How does the size of the droplets affect the rainbow if the diameter of droplets get
close to the wavelength of the light?

Simulating the formation of a primary rainbow due to the refraction and reflection of light within water droplets. We can use the principles of geometric optics and Snell's law to calculate the angles at which different colors appear.
Secondary rainbows are formed by a second reflection inside the water droplet, resulting in a larger angular separation from the primary rainbow. We can modify our model to account for this by adding another reflection and calculating the corresponding deviation angle.
Alexander's Band is a dark band of sky between the primary and secondary rainbows. It occurs because there is less light scattered back to the observer at smaller scattering angles. We can visualize this by plotting the intensity of light as a function of scattering angle.
This code includes the functionalities for calculating primary and secondary rainbows, as well as intensity vs. scattering angle, all in one script